Added middleware to the contacts apis for JWT authentication.

+++ internal/model/contact.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type Contact struct {
-	ID        int       `json:"id"`
-	UserID    int       `json:"user_id"`
-	ContactID int       `json:"contact_id"`
+	ID        int64     `json:"id"`
+	UserID    int64     `json:"user_id"`
+	ContactID int64     `json:"contact_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
